ISSUE #1, #2, #3 - tipologie ricette ha assunto cardinalità N a N - inserito il sistema di gestione delle preferite - iniziato a costruire il form di ricerca

+++ librerie/classi/controller/RicettaController.php
@@ -9,6 +9,8 @@ class RicettaController {
     private $tDAO;
     private $rDAO;
     private $irDAO;
+    private $rtDAO;
+    private $pDAO;
     private $iC;
     
     function __construct() {
@@ -16,6 +18,8 @@ class RicettaController {
         $this->tDAO = new TipologiaDAO();
         $this->rDAO = new RicettaDAO();
         $this->irDAO = new IngredienteRicettaDAO();
+        $this->rtDAO = new RicettaTipologiaDAO();
+        $this->pDAO = new PreferiteDAO();
         //assegno la classe Controller di ingrediente
         $this->iC = new IngredienteController();
     }
@@ -97,7 +101,7 @@ class RicettaController {
                 'formato'   => 'INT'
             )
         );
-        if($this->rDAO->getRicette($query) == null){
+        if($this->rtDAO->getRicetteTipoologie($query) == null){
             return false;
         }
         return true;
@@ -111,8 +115,9 @@ class RicettaController {
      * @param type $arrayIR
      * @return boolean
      */
-    public function saveRicetta(Ricetta $r, $arrayIR){    
-        //arrayIR deve essere un array di oggetti IngredienteRicetta        
+    public function saveRicetta(Ricetta $r, $arrayIR, $arrayT){    
+        //arrayIR deve essere un array di oggetti IngredienteRicetta  
+        //arrayRT deve essere un array di idTipologia      
         //1. Salvo la ricetta
         $idRicetta = $this->rDAO->saveRicetta($r);
         
@@ -125,6 +130,17 @@ class RicettaController {
                 $ir->setIdRicetta($idRicetta);                
                 if($this->irDAO->saveIngredienteRicetta($ir) == false){
                     //cancello la ricetta salvata precedentemente
+                    $this->rDAO->deleteRicettaByID($idRicetta);
+                    return false;
+                }
+            }
+            //associo le tipologie alla ricetta
+            foreach($arrayT as $idTipologia){
+                if($this->rtDAO->saveRicettaTipologia($idRicetta, $idTipologia) == false){
+                    //cancello gli ingredienti associati alla ricetta
+                    $query = array('id_ricetta' => $idRicetta);
+                    $this->irDAO->deleteIngredientiRicette($query);
+                    //cancello la ricetta
                     $this->rDAO->deleteRicettaByID($idRicetta);
                     return false;
                 }
@@ -199,24 +215,7 @@ class RicettaController {
         
         return $this->getRicette($query);        
     }
-    
-    /**
-     * La funzione restituisce tutte le ricette di una determinata tipologia
-     * @param type $idTipologia
-     * @return type
-     */
-    public function getRicetteByTipologia($idTipologia){
-        $query = array(
-            array(
-                'campo'     => 'id_tipologia',
-                'valore'    => $idTipologia,
-                'formato'   => 'INT'
-            )
-        );
-        
-        return $this->getRicette($query);
-    }
-    
+            
     /**
      * La funzione restituisce tutte le ricette dati determinati parametri in ingresso
      * @param type $param
@@ -264,7 +263,21 @@ class RicettaController {
                     //salvo gli ingredienti Ricetta nella ricettta
                     $r->setIngredienti($temp);
                 }
-
+                
+                //salvo le tipologie ricetta nella ricetta
+                $temp2 = $this->rtDAO->getRicetteTipoologie($query);
+                if($temp2 != null){
+                    $tipologie = array();
+                    foreach($temp2 as $item){
+                        
+                        $tp = $this->tDAO->getTipologiaById($item['idTipologia']);
+                        if($tp!=null){
+                            array_push($tipologie, $tp);
+                        }
+                    }
+                    //salvo le tipologie nella ricetta
+                    $r->setTipologie($tipologie);
+                }
                 
                 //salvo la ricetta aggiornata con gli ingredienti, nell'array result
                 array_push($result, $r);
@@ -279,7 +292,7 @@ class RicettaController {
      * @param type $arrayIR
      * @return boolean
      */
-    public function updateRicetta(Ricetta $r, $arrayIR){
+    public function updateRicetta(Ricetta $r, $arrayIR, $arrayT){
         //l'aggiornamento di una ricetta, consiste non solo nell'aggiornare gli attributi dell'elemento Ricetta,
         //ma consiste anche nell'aggiornare la tabella IngredientiRicette a seconda dei nuovi ingredienti aggiunti
         
@@ -288,6 +301,8 @@ class RicettaController {
         //-1 in caso di errore nell'update sulla ricetta
         //-2 in caso di errore nella cancellazione degli ingredienti associati alla ricetta
         //-3 in caso di errore nel salvare un ingrediente nella ricetta
+        //-4 in caso di errore nella cancellazione di tutte le tipologie
+        //-5 in caso di errore nel salvare la tipoolgia nella ricetta
         
         if($this->rDAO->updateRicetta($r) == true){
             //"aggiorno" anche la lista degli ingredienti
@@ -308,7 +323,20 @@ class RicettaController {
             else{
                 //errore nel cancellare gli ingredienti dalla ricetta
                 return -2;
-            }            
+            }
+            //3. elimino tutte le tipologie
+            if($this->rtDAO->deleteRicetteTipologie($query) == true){
+                foreach($arrayT as $idTipologia){
+                    if($this->rtDAO->saveRicettaTipologia($r->getID(), $idTipologia) == false){
+                        return -5;
+                    }
+                }
+            }
+            else{
+                return -4;
+            }
+            
+            
             return true;
         }
         //errore nell'aggiornare la ricetta
@@ -338,14 +366,45 @@ class RicettaController {
      */
     public function deleteRicetta($idRicetta){
         //per eliminare una ricetta, prima devo eliminare la lista di ingredienti associata
+        //update--> devo eliminare anche tutte le preferenze su quella ricetta prima di eliminarla
         $query = array('id_ricetta' => $idRicetta);
         if($this->irDAO->deleteIngredientiRicette($query) == true){
-            //elimino la ricetta
-            if($this->rDAO->deleteRicettaByID($idRicetta) == true){
-                return true;
+            //elimino le preferite sulla ricetta
+            if($this->pDAO->deletePreferite($query) == true){
+                //elimino la ricetta
+                if($this->rDAO->deleteRicettaByID($idRicetta) == true){
+                    return true;
+                }
             }
         }
         return false;
+    }
+    
+    /**
+     * La funzione salva una ricetta come preferita per un determinato utente
+     * @param type $idUtente
+     * @param type $idPreferita
+     * @return boolean
+     */
+    public function savePreferita($idUtente, $idRicetta){
+        if($this->pDAO->savePreferita($idUtente, $idRicetta)==false){
+            return false;
+        }
+        return true;
+    }
+    
+    /**
+     * La funzione elimina un associazione preferita tra un utente e una ricetta
+     * @param type $idUtente
+     * @param type $idRicetta
+     * @return type
+     */
+    public function removePreferita($idUtente, $idRicetta){
+        $query = array(
+            'id_utente'  => $idUtente,
+            'id_ricetta' => $idRicetta 
+        );
+        return $this->pDAO->deletePreferite($query);
     }
